requirements.txt
selenium
pytest
pytest-html
python-dotenv
allure-pytest